+++ Tests/TestContig.cs
@@ -23,7 +23,7 @@ public class TestContig
     [Test]
     public void TestSplit()
     {
-        int remainderLen = _contig1.Length() - 1000;
+        long remainderLen = _contig1.Length() - 1000;
         var rest = _contig1.Split(1000, true);
         Assert.AreEqual(1000, _contig1.Length());
         Assert.AreEqual(remainderLen, rest.Length());
@@ -32,7 +32,7 @@ public class TestContig
     [Test]
     public void TestJoin()
     {
-        int combinedLen = _contig1.Length() + _contigX.Length(); 
+        long combinedLen = _contig1.Length() + _contigX.Length(); 
         _contig1.Join(_contigX);
         Assert.AreEqual(combinedLen, _contig1.Length());
     }
@@ -40,7 +40,7 @@ public class TestContig
     [Test]
     public void TestInversion()
     {
-        int length = _contig1.Length(); 
+        long length = _contig1.Length(); 
         _contig1.InvertRange(length / 4, length * 3 / 4);
         Assert.AreEqual(length, _contig1.Length());
     }
@@ -48,7 +48,7 @@ public class TestContig
     [Test]
     public void TestReplication()
     {
-        int length = _contig1.Length() + 900;
+        long length = _contig1.Length() + 900;
         _contig1.DuplicateRange(100, 1000);
         Assert.AreEqual(length, _contig1.Length());
     }
@@ -56,7 +56,7 @@ public class TestContig
     [Test]
     public void TestBridgeFront()
     {
-        int length = (_contig1.Length() - 1000) * 2;
+        long length = (_contig1.Length() - 1000) * 2;
         _contig1.Bridge(1000, true);
         Assert.AreEqual(length, _contig1.Length());
         _contig1.Bridge(1000, false);
@@ -66,8 +66,8 @@ public class TestContig
     [Test]
     public void TestScatterAndGather()
     {
-        int length = _contig1.Length();
-        _contig1.ScatterAndGather(new List<int>{1000, 2000, 3000}, new List<int>{3, 1, 2, 0});
+        long length = _contig1.Length();
+        _contig1.ScatterAndGather(new List<long>{1000, 2000, 3000}, new List<int>{3, 1, 2, 0});
         Assert.AreEqual(length, _contig1.Length());
     }
 }
